Ensure captcha is completed before sso login
