+++ src/connection/database/login_pane.jsx
@@ -3,7 +3,8 @@ import React from 'react';
 import EmailPane from '../../field/email/email_pane';
 import UsernamePane from '../../field/username/username_pane';
 import PasswordPane from '../../field/password/password_pane';
-import { showResetPasswordActivity, swapCaptcha } from './actions';
+import { showResetPasswordActivity } from './actions';
+import { swapCaptcha } from '../captcha';
 import { hasScreen, forgotPasswordLink } from './index';
 import * as l from '../../core/index';
 import CaptchaPane from '../../field/captcha/captcha_pane';
