+++ src/connection/enterprise/actions.js
@@ -9,6 +9,7 @@ import { getFieldValue, hideInvalidFields } from '../../field/index';
 import { emailLocalPart } from '../../field/email';
 import { logIn as coreLogIn } from '../../core/actions';
 import * as l from '../../core/index';
+import { setCaptchaParams, showMissingCaptcha } from '../captcha';
 
 // TODO: enterprise connections should not depend on database
 // connections. However, we now allow a username input to contain also
@@ -41,10 +42,18 @@ export function logIn(id) {
   const ssoConnection = matchConnection(m, email);
   const enterpriseConnection = enterpriseActiveFlowConnection(m);
   const connectionScopes = getConnectionScopesFrom(m, ssoConnection || enterpriseConnection);
+  const usernameField = databaseLogInWithEmail(m) ? 'email' : 'username';
+  const fields = [usernameField, 'password'];
 
   const params = {
     connection_scope: connectionScopes ? connectionScopes.toJS() : undefined
   };
+
+  const isCaptchaValid = setCaptchaParams(m, params, fields);
+
+  if (!isCaptchaValid) {
+    return showMissingCaptcha(m, id);
+  }
 
   if (ssoConnection && !isHRDActive(m)) {
     return logInSSO(id, ssoConnection, params);
